Create different & working approach for RESTful services

+++ backend/src/main/java/com/example/backend/repository/ComparisonConditionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComparisonConditionRepository extends JpaRepository<ComparisonCondition, Long> {
+public interface ComparisonConditionRepository extends JpaRepository<ComparisonCondition, Integer> {
 }
