+++ backend/src/main/java/com/example/backend/repository/CriteriaTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CriteriaTypeRepository extends JpaRepository<CriteriaType, Long> {
+public interface CriteriaTypeRepository extends JpaRepository<CriteriaType, Integer> {
 }
